libs/hotelstaff_shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hotelstaff-shared"
version = "0.1.0"
description = "Componentes transversales compartidos entre microservicios de HotelStaffIA"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "structlog>=24.4",
    "aio-pika>=9.4",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "redis>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["hotelstaff_shared"]
